+++ CHANGELOG.md
@@ -1,8 +1,17 @@
-## 3.3.2 (YYYY-MM-DD)
+## 3.3.3 (YYYY-MM-DD)
 
 ### Breaking Changes
 
 ### Enhancements
+
+### Bug Fixes
+
+* When converting nullable BLOB field to required, `null` values should be converted to `byte[0]` instead of `byte[1]`.
+* Fixed a bug which may cause duplicated primary key values when migrating a nullable primary key field to not nullable. `RealmObjectSchema.setRequired()` and `RealmObjectSchema.setNullable()` will throw when converting a nullable primary key field with null values stored to a required primary key field.
+
+### Internal
+
+## 3.3.2 (2017-06-09)
 
 ### Bug Fixes